+++ Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
@@ -1,0 +1,23 @@
+using Esprima.Ast;
+using Jint.Runtime.Environments;
+
+namespace Jint.Runtime.Interpreter.Expressions
+{
+    internal sealed class JintIdentifierExpression : JintExpression<Identifier>
+    {
+        private readonly string _expressionName;
+
+        public JintIdentifierExpression(Engine engine, Identifier expression) : base(engine, expression)
+        {
+            _expressionName = _expression.Name;
+        }
+
+        public override object Evaluate()
+        {
+            var env = _engine.ExecutionContext.LexicalEnvironment;
+            var strict = StrictModeScope.IsStrictModeCode;
+
+            return LexicalEnvironment.GetIdentifierReference(env, _expressionName, strict);
+        }
+    }
+}
