+++ Jint/Runtime/Interpreter/Statements/JintBreakStatement.cs
@@ -1,0 +1,22 @@
+using Esprima.Ast;
+
+namespace Jint.Runtime.Interpreter.Statements
+{
+    /// <summary>
+    /// http://www.ecma-international.org/ecma-262/5.1/#sec-12.8
+    /// </summary>
+    internal sealed class JintBreakStatement : JintStatement<BreakStatement>
+    {
+        private readonly string _label;
+
+        public JintBreakStatement(Engine engine, BreakStatement statement) : base(engine, statement)
+        {
+            _label = statement.Label?.Name;
+        }
+
+        public override Completion Execute()
+        {
+            return new Completion(CompletionType.Break, null, _label);
+        }
+    }
+}
